requirements.txt
alembic==1.0.5
amqp==2.3.2
aniso8601==4.0.1
billiard==3.5.0.5
blinker==1.4
celery==4.2.1
Click==7.0
Flask==1.0.2
Flask-DotEnv==0.1.1
Flask-JWT==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
kombu==4.2.1
Mako==1.0.7
MarkupSafe==1.1.0
PyJWT==1.4.2
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
pytz==2018.7
six==1.12.0
SQLAlchemy==1.2.15
sqlparse==0.2.4
sqltap==0.3.10
vine==1.1.4
Werkzeug==0.14.1
